Move the ClusterTopologyReconciler to a separated package
