+++ controllers/topology/util.go
@@ -1,0 +1,94 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package topology
+
+import (
+	"context"
+	"strings"
+
+	"github.com/pkg/errors"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"sigs.k8s.io/cluster-api/controllers/external"
+	utilconversion "sigs.k8s.io/cluster-api/util/conversion"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// getTemplate gets the object referenced in ref.
+// If necessary, it updates the ref to the latest apiVersion of the current contract.
+func (r *ClusterReconciler) getTemplate(ctx context.Context, ref *corev1.ObjectReference) (*unstructured.Unstructured, error) {
+	if ref == nil {
+		return nil, errors.New("reference is not set")
+	}
+	if err := utilconversion.ConvertReferenceAPIContract(ctx, r.Client, r.restConfig, ref); err != nil {
+		return nil, err
+	}
+
+	obj, err := external.Get(ctx, r.UnstructuredCachingClient, ref, ref.Namespace)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to retrieve %s %q in namespace %q", ref.Kind, ref.Name, ref.Namespace)
+	}
+	return obj, nil
+}
+
+// getNestedRef returns the ref value of a nested field.
+func getNestedRef(obj *unstructured.Unstructured, fields ...string) (*corev1.ObjectReference, error) {
+	ref := &corev1.ObjectReference{}
+	if v, ok, err := unstructured.NestedString(obj.UnstructuredContent(), append(fields, "apiVersion")...); ok && err == nil {
+		ref.APIVersion = v
+	} else {
+		return nil, errors.Errorf("failed to get %s.apiVersion from %s", strings.Join(fields, "."), obj.GetKind())
+	}
+	if v, ok, err := unstructured.NestedString(obj.UnstructuredContent(), append(fields, "kind")...); ok && err == nil {
+		ref.Kind = v
+	} else {
+		return nil, errors.Errorf("failed to get %s.kind from %s", strings.Join(fields, "."), obj.GetKind())
+	}
+	if v, ok, err := unstructured.NestedString(obj.UnstructuredContent(), append(fields, "name")...); ok && err == nil {
+		ref.Name = v
+	} else {
+		return nil, errors.Errorf("failed to get %s.name from %s", strings.Join(fields, "."), obj.GetKind())
+	}
+	if v, ok, err := unstructured.NestedString(obj.UnstructuredContent(), append(fields, "namespace")...); ok && err == nil {
+		ref.Namespace = v
+	} else {
+		return nil, errors.Errorf("failed to get %s.namespace from %s", strings.Join(fields, "."), obj.GetKind())
+	}
+	return ref, nil
+}
+
+// setNestedRef sets the value of a nested field to a reference to the refObj provided.
+func setNestedRef(obj, refObj *unstructured.Unstructured, fields ...string) error {
+	ref := map[string]interface{}{
+		"kind":       refObj.GetKind(),
+		"namespace":  refObj.GetNamespace(),
+		"name":       refObj.GetName(),
+		"apiVersion": refObj.GetAPIVersion(),
+	}
+	return unstructured.SetNestedField(obj.UnstructuredContent(), ref, fields...)
+}
+
+// objToRef returns a reference to the given object.
+func objToRef(obj client.Object) *corev1.ObjectReference {
+	gvk := obj.GetObjectKind().GroupVersionKind()
+	return &corev1.ObjectReference{
+		Kind:       gvk.Kind,
+		APIVersion: gvk.GroupVersion().String(),
+		Namespace:  obj.GetNamespace(),
+		Name:       obj.GetName(),
+	}
+}
